fix ptrack

+++ tests/helpers/ptrack_helpers.py
@@ -231,9 +231,9 @@ class ProbackupTest(object):
     def get_md5_per_page_for_fork(self, file, size_in_pages):
         pages_per_segment = {}
         md5_per_page = {}
-        nsegments = int(size_in_pages / 131072)
+        nsegments = size_in_pages / 131072
         if size_in_pages % 131072 != 0:
-            nsegments = nsegments + 1
+            nsegments = int(nsegments) + 1
         # print("Size: {0}".format(size_in_pages))
         # print("Number of segments: {0}".format(nsegments))
